Fix import crash
`yarn init` and `npm init` don't add a `productName` to `package.json`, and when running import on a project without a `productName`, the config setup would throw an exception, so guard against that. This also implements a very basic import unit test -- maybe we want more, but this at least covers this bug fix and other sanity-check-level issues.